Improving BeeUpgradeModule
Bee will now only delete the destination files that it has in its
source.

Without this change, Bee could delete the basefolder where it was
located which could lead to some problems depending on where it was
installed.

Also there's a piece of code that deletes old jar files of Bee.